Renamed to bedrock.py, updated parameter parsing, system message extraction, client class incorporation
