+++ autogen/logger/sqlite_logger.py
@@ -19,6 +19,7 @@ from .base_logger import LLMConfig
 if TYPE_CHECKING:
     from autogen import Agent, ConversableAgent, OpenAIWrapper
     from autogen.oai.anthropic import AnthropicClient
+    from autogen.oai.bedrock import BedrockClient
     from autogen.oai.cohere import CohereClient
     from autogen.oai.gemini import GeminiClient
     from autogen.oai.groq import GroqClient
@@ -402,6 +403,7 @@ class SqliteLogger(BaseLogger):
             TogetherClient,
             GroqClient,
             CohereClient,
+            BedrockClient,
         ],
         wrapper: OpenAIWrapper,
         init_args: Dict[str, Any],
