+++ autogen/oai/client.py
@@ -84,6 +84,13 @@ try:
 except ImportError as e:
     cohere_import_exception = e
 
+try:
+    from autogen.oai.bedrock import BedrockClient
+
+    bedrock_import_exception: Optional[ImportError] = None
+except ImportError as e:
+    bedrock_import_exception = e
+
 logger = logging.getLogger(__name__)
 if not logger.handlers:
     # Add the console handler.
@@ -518,6 +525,11 @@ class OpenAIWrapper:
                 if cohere_import_exception:
                     raise ImportError("Please install `cohere` to use the Groq API.")
                 client = CohereClient(**openai_config)
+                self._clients.append(client)
+            elif api_type is not None and api_type.startswith("bedrock"):
+                if bedrock_import_exception:
+                    raise ImportError("Please install `boto3` to use the Amazon Bedrock API.")
+                client = BedrockClient(**openai_config)
                 self._clients.append(client)
             else:
                 client = OpenAI(**openai_config)
