+++ autogen/runtime_logging.py
@@ -14,6 +14,7 @@ from autogen.logger.logger_factory import LoggerFactory
 if TYPE_CHECKING:
     from autogen import Agent, ConversableAgent, OpenAIWrapper
     from autogen.oai.anthropic import AnthropicClient
+    from autogen.oai.bedrock import BedrockClient
     from autogen.oai.cohere import CohereClient
     from autogen.oai.gemini import GeminiClient
     from autogen.oai.groq import GroqClient
@@ -113,7 +114,15 @@ def log_new_wrapper(wrapper: OpenAIWrapper, init_args: Dict[str, Union[LLMConfig
 
 def log_new_client(
     client: Union[
-        AzureOpenAI, OpenAI, GeminiClient, AnthropicClient, MistralAIClient, TogetherClient, GroqClient, CohereClient
+        AzureOpenAI,
+        OpenAI,
+        GeminiClient,
+        AnthropicClient,
+        MistralAIClient,
+        TogetherClient,
+        GroqClient,
+        CohereClient,
+        BedrockClient,
     ],
     wrapper: OpenAIWrapper,
     init_args: Dict[str, Any],
